requirements.txt
###### Requirements with Version Specifiers ######`

numpy == 1.18.1
matplotlib == 3.1.3
scikit-learn == 0.22.1
GitPython == 3.0.5


